Extract getDefaultAxisRange function
